// shopify.d.ts
export interface RestrictionsObject {
    reason: string;
}
// ...
export interface UsersQueueResponse {
    currently_playing: TrackObjectFull | EpisodeObjectFull;
    queue: Array<TrackObjectFull | EpisodeObjectFull>;
}
// ...
export interface AlbumObjectSimplified extends ContextObject {
    /**
     * The field is present when getting an artist’s albums.
     * Possible values are “album”, “single”, “compilation”, “appears_on”.
     * Compare to album_type this field represents relationship between the artist and the album.
     */
    album_group?: 'album' | 'single' | 'compilation' | 'appears_on' | undefined;
    /**
     * The type of the album: one of “album”, “single”, or “compilation”.
     */
    album_type: 'album' | 'single' | 'compilation';
    /**
     * The artists of the album.
     * Each artist object includes a link in href to more detailed information about the artist.
     */
    artists: ArtistObjectSimplified[];
    /**
     * The markets in which the album is available: [ISO 3166-1 alpha-2 country codes](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2).
     * Note that an album is considered available in a market when at least 1 of its tracks is available in that market.
     */
    available_markets?: string[] | undefined;
    /**
     * The [Spotify ID](https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids) for the album.
     */
    id: string;
    /**
     * The cover art for the album in various sizes, widest first.
     */
    images: ImageObject[];
    /**
     * The name of the album. In case of an album takedown, the value may be an empty string.
     */
    name: string;
    /**
     * The date the album was first released, for example `1981`.
     * Depending on the precision, it might be shown as `1981-12` or `1981-12-15`.
     */
    release_date: string;
    /**
     * The precision with which release_date value is known: `year`, `month`, or `day`.
     */
    release_date_precision: 'year' | 'month' | 'day';
    /**
     * Part of the response when [Track Relinking](https://developer.spotify.com/documentation/general/guides/track-relinking-guide/) is applied,
     * the original track is not available in the given market, and Spotify did not have any tracks to relink it with.
     * The track response will still contain metadata for the original track,
     * and a restrictions object containing the reason why the track is not available: `"restrictions" : {"reason" : "market"}`
     */
    restrictions?: RestrictionsObject | undefined;
    type: 'album';
    /**
     * The number of tracks in the album.
     */
    total_tracks: number;
}
// ...
export interface CopyrightObject {
    text: string;
    type: 'C' | 'P';
}
// ...
export interface ExternalIdObject {
    isrc?: string | undefined;
    ean?: string | undefined;
    upc?: string | undefined;
}

/**
 * External Url Object
 * [](https://developer.spotify.com/web-api/object-model/)
 *
 * Note that there might be other types available, it couldn't be found in the docs.
 */
export interface ExternalUrlObject {
    spotify: string;
}
// ...
export interface ImageObject {
    /**
     * The image height in pixels. If unknown: `null` or not returned.
     */
    height?: number | undefined;
    /**
     * The source URL of the image.
     */
    url: string;
    /**
     * The image width in pixels. If unknown: null or not returned.
     */
    width?: number | undefined;
}
// ...
export interface TrackObjectFull extends TrackObjectSimplified {
    /**
     * The album on which the track appears.
     */
    album: AlbumObjectSimplified;
    /**
     * Known external IDs for the track.
     */
    external_ids: ExternalIdObject;
    /**
     * The popularity of the track. The value will be between `0` and `100`, with `100` being the most popular.
     * The popularity of a track is a value between `0` and `100`, with `100` being the most popular.
     * The popularity is calculated by algorithm and is based, in the most part,
     * on the total number of plays the track has had and how recent those plays are.
     */
    popularity: number;
    /**
     * Whether or not the track is from a local file.
     */
    is_local?: boolean | undefined;
}

/**
 * Simplified Track Object
 * [track object (simplified)](https://developer.spotify.com/web-api/object-model/#track-object-simplified)
 */
export interface TrackObjectSimplified {
    /**
     * The artists who performed the track.
     */
    artists: ArtistObjectSimplified[];
    /**
     * A list of the countries in which the track can be played,
     * identified by their [ISO 3166-1 alpha-2 code](http://en.wikipedia.org/wiki/ISO_3166-1_alpha-2).
     */
    available_markets?: string[] | undefined;
    /**
     * The disc number (usually `1` unless the album consists of more than one disc).
     */
    disc_number: number;
    /**
     * The track length in milliseconds.
     */
    duration_ms: number;
    /**
     * Whether or not the track has explicit lyrics (`true` = yes it does; `false` = no it does not OR unknown).
     */
    explicit: boolean;
    /**
     * Known external URLs for this track.
     */
    external_urls: ExternalUrlObject;
    /**
     * A link to the Web API endpoint providing full details of the track.
     */
    href: string;
    /**
     * The [Spotify ID](https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids) for the track.
     */
    id: string;
    /**
     * Part of the response when [Track Relinking](https://developer.spotify.com/documentation/general/guides/track-relinking-guide/) is applied.
     * If `true`, the track is playable in the given market. Otherwise, `false`.
     */
    is_playable?: boolean | undefined;
    /**
     * Part of the response when [Track Relinking](https://developer.spotify.com/documentation/general/guides/track-relinking-guide/) is applied,
     * and the requested track has been replaced with different track.
     * The track in the `linked_from` object contains information about the originally requested track.
     */
    linked_from?: TrackLinkObject | undefined;
    /**
     * Part of the response when [Track Relinking](https://developer.spotify.com/documentation/general/guides/track-relinking-guide/) is applied,
     * the original track is not available in the given market, and Spotify did not have any tracks to relink it with.
     * The track response will still contain metadata for the original track, and a restrictions object containing the reason
     * why the track is not available: `"restrictions" : {"reason" : "market"}`.
     */
    restrictions?: RestrictionsObject | undefined;
    /**
     * The name of the track.
     */
    name: string;
    /**
     * A link to a 30 second preview (MP3 format) of the track. Can be null
     */
    preview_url: string | null;
    /**
     * The number of the track. If an album has several discs, the track number is the number on the specified disc.
     */
    track_number: number;
    /**
     * The object type: “track”.
     */
    type: 'track';
    /**
     * The [Spotify URI](https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids) for the track.
     */
    uri: string;
}

/**
 * Track Link Object
 * [](https://developer.spotify.com/web-api/object-model/#track-object-simplified)
 */
export interface TrackLinkObject {
    external_urls: ExternalUrlObject;
    href: string;
    id: string;
    type: 'track';
    uri: string;
}

/**
 * Episode Object
 * [episode object](https://developer.spotify.com/documentation/web-api/reference/#object-episodeobject)
 */
export interface EpisodeObject extends EpisodeObjectSimplified {
    /**
     * The show on which the episode belongs.
     */
    show: ShowObjectSimplified;
    /**
     * The [Spotify URI](https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids) for the episode.
     */
    uri: string;
}

export interface EpisodeObjectFull extends EpisodeObject {}

/**
 * Simplified Episode Object
 * [episode object (simplified)](https://developer.spotify.com/documentation/web-api/reference/object-model/#episode-object-simplified)
 */
export interface EpisodeObjectSimplified extends ContextObject {
    /**
     * A URL to a 30 second preview (MP3 format) of the episode. null if not available.
     */
    audio_preview_url: string | null;
    /**
     * A description of the episode.
     */
    description: string;
    /**
     * The episode length in milliseconds.
     */
    duration_ms: number;
    /**
     * Whether or not the episode has explicit content (true = yes it does; false = no it does not OR unknown).
     */
    explicit: boolean;
    /**
     * A description of the episode. This field may contain HTML tags.
     */
    html_description: string;
    /**
     * The [Spotify ID](https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids) for the episode.
     */
    id: string;
    /**
     * The cover art for the episode in various sizes, widest first.
     */
    images: ImageObject[];
    /**
     * True if the episode is hosted outside of Spotify’s CDN.
     */
    is_externally_hosted: boolean;
    /**
     * True if the episode is playable in the given market. Otherwise false.
     */
    is_playable: boolean;
    /**
     * The language used in the episode, identified by a [ISO 639](https://en.wikipedia.org/wiki/ISO_639) code.
     * @deprecated Note: This field is deprecated and might be removed in the future. Please use the languages field instead.
     */
    language: string;
    /**
     * A list of the languages used in the episode, identified by their [ISO 639](https://en.wikipedia.org/wiki/ISO_639) code.
     * Optional because sometimes only the deprecated language field is set and this one isn't set at all.
     */
    languages?: string[] | undefined;
    /**
     * The name of the episode.
     */
    name: string;
    /**
     * The date the episode was first released, for example "1981-12-15". Depending on the precision, it might be shown as "1981" or "1981-12".
     */
    release_date: string;
    /**
     * The precision with which release_date value is known: "year", "month", or "day".
     */
    release_date_precision: string;
    /**
     * The user’s most recent position in the episode. Set if the supplied access token is a user token and has the scope user-read-playback-position.
     */
    resume_point?: ResumePointObject | undefined;
    type: 'episode';
}
// ...
export interface ShowObjectSimplified extends ContextObject {
    /**
     * A list of the countries in which the show can be played, identified by their [ISO 3166-1 alpha-2 code](http://en.wikipedia.org/wiki/ISO_3166-1_alpha-2).
     */
    available_markets: string[];
    /**
     * The copyright statements of the show.
     */
    copyrights: CopyrightObject[];
    /**
     * A description of the show.
     */
    description: string;
    /**
     * A description of the show. This field may contain HTML tags.
     */
    html_description: string;
    /**
     * Whether or not the show has explicit content (true = yes it does; false = no it does not OR unknown).
     */
    explicit: boolean;
    /**
     * The [Spotify ID](https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids) for the show.
     */
    id: string;
    /**
     * The cover art for the show in various sizes, widest first.
     */
    images: ImageObject[];
    /**
     * True if all of the show’s episodes are hosted outside of Spotify’s CDN. This field might be null in some cases.
     */
    is_externally_hosted: boolean | null;
    /**
     * A list of the languages used in the show, identified by their [ISO 639](https://en.wikipedia.org/wiki/ISO_639) code.
     */
    languages: string[];
    /**
     * The media type of the show.
     */
    media_type: string;
    /**
     * The name of the show.
     */
    name: string;
    /**
     * The publisher of the show.
     */
    publisher: string;
    /**
     * The object type: “show”.
     */
    type: 'show';
    // This is found in https://developer.spotify.com/documentation/web-api/reference/shows/get-a-show/ but not in
    // https://developer.spotify.com/documentation/web-api/reference/object-model/#show-object-full.
    // Also it is not always sent, so it is marked optional here.
    /**
     * Total number of episodes in the show.
     */
    total_episodes?: number | undefined;
}
// ...
export interface ContextObject {
    /**
     * The object type.
     */
    type: 'artist' | 'playlist' | 'album' | 'show' | 'episode';
    /**
     * A link to the Web API endpoint providing full details.
     */
    href: string;
    /**
     * Known external URLs.
     */
    external_urls: ExternalUrlObject;
    /**
     * The [Spotify URI](https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids).
     */
    uri: string;
}